% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{additionalColumns}
\alias{additionalColumns}
\title{Identify variables in additional_name column}
\usage{
additionalColumns(result)
}
\arguments{
\item{result}{A tibble.}
}
\value{
Unique values of the additional name column.
}
\description{
Identifies and returns the unique values in additional_name
column.
}
\examples{
mockSummarisedResult() |>
  additionalColumns()

}
