% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{alluvialPlot}
\alias{alluvialPlot}
\title{Create an alluvial plot visualisation from a data frame or a
\verb{<summarised_result>} object.}
\usage{
alluvialPlot(result, x, y, colour = x, facet = NULL, style = NULL, type = NULL)
}
\arguments{
\item{result}{A \verb{<summarised_result>} object.}

\item{x}{A character vector of column names to use as alluvial axes, in
order from left to right. Must contain at least 2 elements.}

\item{y}{Column or estimate name that is used as y variable.}

\item{colour}{Columns to use to determine the colours.}

\item{facet}{Variables to facet by, a formula can be provided to specify
which variables should be used as rows and which ones as columns.}

\item{style}{Visual theme to apply. Character, or \code{NULL}.
If a character, this may be either the name of a built-in style
(see \code{plotStyle()}), or a path to a \code{.yml} file that
defines a custom style. If \code{NULL}, the function will use the
explicit default style, unless a global style option is
set (see \code{setGlobalPlotOptions()}), or a \verb{_brand.yml} file is present
(in that order).
Refer to the package vignette on styles to learn more.}

\item{type}{Character string indicating the output plot format.
See \code{plotType()} for the list of supported plot types. If \code{type = NULL},
the function will use the global setting defined via \code{setGlobalPlotOptions()}
(if available); otherwise, a standard \code{ggplot2} plot is produced by default.}
}
\value{
A plot object.
}
\description{
Create an alluvial plot visualisation from a data frame or a
\verb{<summarised_result>} object.
}
\examples{
result <- dplyr::tibble(
  treatment_1 = c("A", "A", "A", "B", "B", "B", "C", "C"),
  treatment_2 = c("A", "A", "B", "A", "B", "B", "B", "C"),
  treatment_3 = c("A", "B", "B", "A", "A", "B", "B", "C"),
  count       = c(22, 3, 5, 7, 3, 17, 4, 12)
)

# basic alluvial plot with 3 axes
alluvialPlot(
  result = result,
  x = c("treatment_1", "treatment_2", "treatment_3"),
  y = "count"
)

# colour by first axis
alluvialPlot(
  result = result,
  x = c("treatment_1", "treatment_2", "treatment_3"),
  y = "count",
  colour = "treatment_1"
)

# colour by multiple variables
alluvialPlot(
  result = result,
  x = c("treatment_1", "treatment_2", "treatment_3"),
  y = "count",
  colour = c("treatment_1", "treatment_2")
)
}
