% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendSettings.R
\name{appendSettings}
\alias{appendSettings}
\title{Append settings as rows in a summarised result}
\usage{
appendSettings(x, colsSettings)
}
\arguments{
\item{x}{A tibble with columns from summarised result and columns
corresponding to settings.}

\item{colsSettings}{Columns of x to append as settings rows.}
}
\value{
A tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
result <- mockSummarisedResult()[1,] |>
dplyr::mutate(
  mock_default = TRUE,
  example_setting = 1
)
appendSettings(result, colsSettings = c("mock_default", "example_setting"))

}
