% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockResults.R
\docType{data}
\name{data}
\alias{data}
\title{List of mock results}
\format{
A list of mock results for quarto and shiny vignette examples
}
\usage{
data
}
\description{
List of mock results
}
\keyword{datasets}
