% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{emptyPlot}
\alias{emptyPlot}
\title{Returns an empty plot}
\usage{
emptyPlot(title = "No data to plot", subtitle = "")
}
\arguments{
\item{title}{Title to use in the empty plot.}

\item{subtitle}{Subtitle to use in the empty plot.}
}
\value{
An empty ggplot object
}
\description{
Returns an empty plot
}
\examples{
emptyPlot()

}
