% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visTable.R
\name{emptyTable}
\alias{emptyTable}
\title{Returns an empty table}
\usage{
emptyTable(type = "gt")
}
\arguments{
\item{type}{The desired format of the output table. See \code{tableType()} for
allowed options.}
}
\value{
An empty table of the class specified in \code{type}
}
\description{
Returns an empty table
}
\examples{
emptyTable(type = "flextable")

}
