% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTable.R
\name{flextableStyle}
\alias{flextableStyle}
\title{Default style code expresion for flextable tables.}
\usage{
flextableStyle(nameStyle = "default")
}
\arguments{
\item{styleName}{Name of the style. Currently the package just have one
predefined style ("default").}
}
\description{
Default style code expresion for flextable tables.
}
