% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatEstimateValue.R
\name{formatEstimateValue}
\alias{formatEstimateValue}
\title{Format estimate_value column of summarised_result and compared_result
object.}
\usage{
formatEstimateValue(
  result,
  decimals = c(integer = 0, numeric = 2, percentage = 1, proportion = 3),
  decimalMark = ".",
  thousandsMark = ","
)
}
\arguments{
\item{result}{A summarised_result or compared_result.}

\item{decimals}{Number of decimals per estimate_type.}

\item{decimalMark}{Decimal separator mark.}

\item{thousandsMark}{Thousand and millions separator mark.}
}
\description{
Format estimate_value column of summarised_result and compared_result
object.
}
\examples{
\donttest{
x <- 1
}

}
