% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatEstimateValue.R
\name{formatMinCellCount}
\alias{formatMinCellCount}
\title{To indicate which was the minimum cell counts where estimates have been
suppressed.}
\usage{
formatMinCellCount(result)
}
\arguments{
\item{result}{A \verb{<summarised_result>} object.}
}
\description{
To indicate which was the minimum cell counts where estimates have been
suppressed.
}
\examples{
result <- mockSummarisedResult()
result |> formatMinCellCount()
}
