% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtSummarisedResult.R
\name{formatSummarisedResult}
\alias{formatSummarisedResult}
\title{Create a gt table from a summarised_result object.}
\usage{
formatSummarisedResult(
  result,
  format = NULL,
  splitGroup = TRUE,
  splitStrata = TRUE,
  splitAdditional = FALSE,
  columns = c("variable_name", "variable_level", format = "estimate_name"),
  header = c("cdm_name", "group", "strata"),
  keepNotFormatted = TRUE,
  decimals = c(integer = 0, numeric = 2, percentage = 1, proportion = 3),
  decimalMark = ".",
  bigMark = ",",
  style = list(header = list(gt::cell_fill(color = "#c8c8c8"), gt::cell_text(weight =
    "bold")), header_name = list(gt::cell_fill(color = "#e1e1e1")), header_level =
    list(gt::cell_fill(color = "#ffffff"))),
  na = "-"
)
}
\arguments{
\item{result}{A summarised_result object.}

\item{format}{#To inherit from formatEstimateName}

\item{splitGroup}{Whether to split group_name and group_level in several
columns.}

\item{splitStrata}{Whether to split strata_name and strata_level in several
columns.}

\item{splitAdditional}{Whether to split additional_name and additional_level
in several columns.}

\item{columns}{Columns to keep. You can rename them with this command.}

\item{header}{Columns to use as spanners in the gt table.}

\item{keepNotFormatted}{#To inherit formatEstimateName}

\item{decimals}{#To inherit formatEstimateValue}

\item{decimalMark}{#To inherit formatEstimateValue}

\item{bigMark}{#To inherit formatEstimateValue}

\item{style}{#To inherit gtHeader}

\item{na}{Substitute for NA values.}
}
\value{
A formatted gt table.
}
\description{
Create a gt table from a summarised_result object.
}
