% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTable.R
\name{formatTable}
\alias{formatTable}
\title{Format a table into a gt, flextable or tibble object.}
\usage{
formatTable(
  result,
  formatEstimateName = character(),
  header = character(),
  groupColumn = character(),
  type = "gt",
  renameColumns = character(),
  hide = character(),
  .options = list()
)
}
\arguments{
\item{result}{A summarised_result.}

\item{formatEstimateName}{Named list of estimate name's to join, sorted by
computation order. Indicate estimate_name's between <...>.}

\item{groupColumn}{Columns to use as group labels. By default the name of the
new group will be the column names separated by "_". To specify a new
grouping name enter a named list such as:
list(newGroupName = c("variable_name", "variable_level"))}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{renameColumns}{Named vector to customise column names, for instance:
c("Database name" = "cdm_name")). By default, "_" are removed from column
names and converted to sentence case.}

\item{hide}{Columns to drop from the output table.}

\item{.options}{Named list with additional formatting options.
visOmopResults::optionsTable() shows allowed arguments and
their default values.}
}
\value{
A tibble, gt, or flextable object.
}
\description{
This function combines the functionalities of \code{formatEstimateValue()},
\code{formatEstimateName()}, \code{formatHeader()}, \code{gtTable()}, and \code{fxTable()}
into a single function.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
