% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTable.R
\name{fxTable}
\alias{fxTable}
\title{Creats a flextable from a dataframe using as deliminter (\code{delim}) for spanners.}
\usage{
fxTable(
  x,
  delim = "\\n",
  style = list(header = list(cell = officer::fp_cell(background.color = "#c8c8c8"), text
    = officer::fp_text(bold = TRUE)), header_name = list(cell =
    officer::fp_cell(background.color = "#d9d9d9"), text = officer::fp_text(bold =
    TRUE)), header_level = list(cell = officer::fp_cell(background.color = "#e1e1e1"),
    text = officer::fp_text(bold = TRUE)), column_name = list(text =
    officer::fp_text(bold = TRUE)), group_label = list(cell =
    officer::fp_cell(background.color = "#e9ecef", border = officer::fp_border(width = 1,
    
     color = "gray")), text = officer::fp_text(bold = TRUE))),
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupNameCol = NULL,
  groupNameAsColumn = FALSE,
  groupOrder = NULL
)
}
\arguments{
\item{x}{A dataframe.}

\item{delim}{Delimiter.}

\item{style}{Named list that specifies how to style the different parts of
the table. Accepted entries are: title, subtitle, header, header_name,
header_level, column_name, group_label, and body.}

\item{na}{How to display missing values.}

\item{title}{Title of the table, or NULL for no title.}

\item{subtitle}{Subtitle of the table, or NULL for no subtitle.}

\item{caption}{Caption for the table, or NULL for no caption. Text in
markdown formatting style (e.g. \verb{*Your caption here*} for caption in
italics).}

\item{groupNameCol}{Column to use as group labels.}

\item{groupNameAsColumn}{Whether to display the group labels as a column
(TRUE) or rows (FALSE).}

\item{groupOrder}{Order in which to display group labels.}
}
\value{
A flextable.
}
\description{
Creats a flextable from a dataframe using as deliminter (\code{delim}) for spanners.
}
\examples{
\donttest{
mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  spanHeader(header = c("Study strata", "strata_level"),
             includeHeader = FALSE) |>
  fxTable(
    style = list(
      "header" = list(
        "cell" = officer::fp_cell(background.color = "#c8c8c8"),
        "text" = officer::fp_text(bold = TRUE)),
      "header_level" = list(
        "cell" = officer::fp_cell(background.color = "#e1e1e1"),
        "text" = officer::fp_text(bold = TRUE)),
      "group_label" = list(
        "cell" = officer::fp_cell(
          background.color = "#e9ecef",
          border = officer::fp_border(width = 1, color = "gray")),
        "text" = officer::fp_text(bold = TRUE))
    ),
    na = "--",
   title = "fxTable example",
   subtitle = NULL,
   caption = NULL,
  groupNameCol = "group_level",
   groupNameAsColumn = TRUE,
   groupOrder = c("cohort1", "cohort2")
 )
}

}
