% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{groupColumns}
\alias{groupColumns}
\title{Identify variables in group_name column}
\usage{
groupColumns(result)
}
\arguments{
\item{result}{A tibble.}
}
\value{
Unique values of the group name column.
}
\description{
Identifies and returns the unique values in group_name column.
}
\examples{
mockSummarisedResult() |>
  groupColumns()

}
