% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtHeader.R
\name{gtHeader}
\alias{gtHeader}
\title{Add headers to a omop result_object.}
\usage{
gtHeader(
  result,
  header = character(),
  style = list(header = list(gt::cell_fill(color = "#c8c8c8"), gt::cell_text(weight =
    "bold")), header_name = list(gt::cell_fill(color = "#e1e1e1")), header_level =
    list(gt::cell_fill(color = "#ffffff"))),
  na = "-"
)
}
\arguments{
\item{result}{A result object.}

\item{header}{Names of the headers. If they point to a column, that column is
used as header.}

\item{style}{Set style for the different headers. Use the labels header (a
header name), header_name () and header_level (). If name is not provided
\code{header} is assumed.}

\item{na}{Substitute for NA.}
}
\description{
Add headers to a omop result_object.
}
