% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtTable.R
\name{gtStyle}
\alias{gtStyle}
\title{Default style code expresion for gt tables.}
\usage{
gtStyle(styleName = "default")
}
\arguments{
\item{styleName}{Name of the style. Currently the package just have one
predefined style ("default").}
}
\description{
Default style code expresion for gt tables.
}
