% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockResults.R
\name{mockSummarisedResult}
\alias{mockSummarisedResult}
\title{A summarised_result mock object.}
\usage{
mockSummarisedResult()
}
\description{
A summarised_result mock object.
}
\examples{
\donttest{
mockSummarisedResult()
}

}
