% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperTableFunctions.R
\name{optionsVisOmopTable}
\alias{optionsVisOmopTable}
\title{Additional table formatting options}
\usage{
optionsVisOmopTable()
}
\value{
list of options
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
