% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotEstimates.R
\name{pivotLongerType}
\alias{pivotLongerType}
\title{Pivot longer multiple columns using the type column, so they can be pivoted
back. All values will be converted to characters.}
\usage{
pivotLongerType(result, cols, prefix = "estimate")
}
\arguments{
\item{result}{A data.frame.}

\item{cols}{Columns of this data.frame that you want to pivot.}

\item{prefix}{Prefix for the new columns.}
}
\value{
The data.frame with the new 3 columns
}
\description{
Pivot longer multiple columns using the type column, so they can be pivoted
back. All values will be converted to characters.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

tibble(
  cdm_name = c("mock 1", "mock 1", "mock 2", "mock 2"),
  cohort_name = c("cohort1", "cohort2", "cohort1", "cohort2"),
  age_mean = c(40L, 32L, 44L, 67L),
  blood_mode = c("A", "B", "0", "B")
) |>
  pivotLongerType(cols = c("age_mean", "blood_mode"))

}
