% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotSettings.R
\name{pivotSettings}
\alias{pivotSettings}
\title{Add settings as columns}
\usage{
pivotSettings(x)
}
\arguments{
\item{x}{A summarised_result.}
}
\value{
A tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Pivot settings rows into columns.
}
\examples{
result <- mockSummarisedResult(settings = TRUE)
result |> pivotSettings()

}
