% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBarplot}
\alias{plotBarplot}
\title{Create a plot visualisation from a summarised result object.}
\usage{
plotBarplot(
  result,
  x,
  y,
  ymax = NULL,
  ymin = NULL,
  facet = NULL,
  colour = NULL
)
}
\arguments{
\item{result}{A summarised result object.}

\item{x}{Column or estimate name that is used as x variable.}

\item{y}{Column or estimate name that is used as y variable}

\item{ymax}{Upper limit of error bars, if provided is plot using
\code{geom_errorbar}.}

\item{ymin}{Lower limit of error bars, if provided is plot using
\code{geom_errorbar}.}

\item{facet}{Variables to facet by, a formula can be provided to specify
which variables should be used as rows and which ones as columns.}

\item{colour}{Columns to use to determine the colors.}
}
\value{
A plot object.
}
\description{
Create a plot visualisation from a summarised result object.
}
\examples{
\donttest{
result <- mockSummarisedResult() |> dplyr::filter(variable_name == "age")

plotBarplot(
  result = result,
  x = "cohort_name",
  y = "mean",
  facet = c("age_group", "sex"))
}

}
