% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{Create a box plot visualisation from a summarised_result object.}
\usage{
plotBoxplot(
  result,
  x = NULL,
  lower = "q25",
  middle = "median",
  upper = "q75",
  ymin = "min",
  ymax = "max",
  facet = NULL,
  colour = NULL
)
}
\arguments{
\item{result}{A summarised result object.}

\item{x}{Columns to use as x axes.}

\item{lower}{Estimate name for the lower quantile of the box.}

\item{middle}{Estimate name for the middle line of the box.}

\item{upper}{Estimate name for the upper quantile of the box.}

\item{ymin}{Estimate name for the lower limit of the bars.}

\item{ymax}{Estimate name for the upper limit of the bars.}

\item{facet}{Variables to facet by, a formula can be provided to specify
which variables should be used as rows and which ones as columns.}

\item{colour}{Columns to use to determine the colors.}
}
\value{
A ggplot2 object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
