% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotColumns}
\alias{plotColumns}
\title{Columns for the plot functions}
\usage{
plotColumns(result)
}
\arguments{
\item{result}{A \verb{<summarised_result>} object.}
}
\value{
A character vector of supported columns for plots.
}
\description{
Names of the columns that can be used in the input arguments for the plot
functions.
}
\examples{
result <- mockSummarisedResult()
plotColumns(result)

}
