% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotfunction}
\alias{plotfunction}
\title{create a ggplot from the output of summarise result.}
\usage{
plotfunction(
  data,
  xAxis = "variable_name",
  yAxis = "estimate_value",
  plotStyle = "scatterplot",
  facetVarX = "variable_name",
  facetVarY = c("group_level", "strata_level"),
  colorVars = "variable_level",
  vertical_x = FALSE
)
}
\arguments{
\item{data}{output of summariseLargeScaleCharacteristics.}

\item{xAxis}{what to plot on x axis, default as variable_name column.
Has to be a column in data.}

\item{yAxis}{what to plot on y axis, default as estimate_value column.
Has to be a column in data. One of the xAxis or yAxis has to be estimate_value.}

\item{plotStyle}{one of barplot, boxplot, scatterplot or density}

\item{facetVarX}{column in data to facet by on horizontal axis}

\item{facetVarY}{column in data to facet by on vertical axis}

\item{colorVars}{column in data to color by.}

\item{vertical_x}{whether to display x axis string vertically.}
}
\value{
A ggplot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\donttest{
library(visOmopResults)
library(dplyr)
result <- mockSummarisedResult() |>
filter(variable_name == "Medications",
strata_name == "overall")
graphs <- plotfunction(result, plotStyle = "barplot")
}
}
