% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotScatter}
\alias{plotScatter}
\title{Create a plot visualisation from a summarised result object.}
\usage{
plotScatter(
  result,
  x,
  y,
  line = TRUE,
  points = TRUE,
  ymin = NULL,
  ymax = NULL,
  facet = NULL,
  colour = NULL,
  group = colour
)
}
\arguments{
\item{result}{A summarised result object.}

\item{x}{Column or estimate name that is used as x variable.}

\item{y}{Column or estimate name that is used as y variable}

\item{line}{Whether to plot a line using \code{geom_line}.}

\item{points}{Whether to plot points using \code{geom_point}.}

\item{ymin}{Lower limit of error bars, if provided is plot using
\code{geom_errorbar}.}

\item{ymax}{Upper limit of error bars, if provided is plot using
\code{geom_errorbar}.}

\item{facet}{Variables to facet by, a formula can be provided to specify
which variables should be used as rows and which ones as columns.}

\item{colour}{Columns to use to determine the colors.}

\item{group}{Columns to use to determine the group.}
}
\value{
A plot object.
}
\description{
Create a plot visualisation from a summarised result object.
}
\examples{
\donttest{
result <- mockSummarisedResult() |>
  dplyr::filter(variable_name == "age")

plotScatter(
  result = result,
  x = "cohort_name",
  y = "mean",
  facet = age_group ~ sex)
}

}
