% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotStyle}
\alias{plotStyle}
\title{Pre-defined styles are available for plots}
\usage{
plotStyle()
}
\value{
A character vector indicating the style names.
}
\description{
This function provides a list of pre-defined styles for plots that can then
be used in the \code{style} argument of the plot functions.
}
\examples{
plotStyle()

}
