% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotType}
\alias{plotType}
\title{Supported plot types}
\usage{
plotType()
}
\value{
A character vector of supported plot types.
}
\description{
This function returns the supported plot types that can be used in the
\code{type} argument of the plot functions.
}
\examples{
tableType()

}
