% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{settings}
\alias{bind}
\alias{exportSummarisedResult}
\alias{importSummarisedResult}
\alias{suppress}
\alias{splitAdditional}
\alias{splitGroup}
\alias{splitStrata}
\alias{splitAll}
\alias{tidyColumns}
\alias{strataColumns}
\alias{additionalColumns}
\alias{groupColumns}
\alias{settingsColumns}
\alias{addSettings}
\alias{pivotEstimates}
\alias{filterAdditional}
\alias{filterGroup}
\alias{filterSettings}
\alias{filterStrata}
\alias{uniteAdditional}
\alias{uniteGroup}
\alias{uniteStrata}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:tidy]{tidy()}}}

  \item{omopgenerics}{\code{\link[omopgenerics:addSettings]{addSettings()}}, \code{\link[omopgenerics:additionalColumns]{additionalColumns()}}, \code{\link[omopgenerics:bind]{bind()}}, \code{\link[omopgenerics:exportSummarisedResult]{exportSummarisedResult()}}, \code{\link[omopgenerics:filterAdditional]{filterAdditional()}}, \code{\link[omopgenerics:filterGroup]{filterGroup()}}, \code{\link[omopgenerics:filterSettings]{filterSettings()}}, \code{\link[omopgenerics:filterStrata]{filterStrata()}}, \code{\link[omopgenerics:groupColumns]{groupColumns()}}, \code{\link[omopgenerics:importSummarisedResult]{importSummarisedResult()}}, \code{\link[omopgenerics:pivotEstimates]{pivotEstimates()}}, \code{\link[omopgenerics:settings]{settings()}}, \code{\link[omopgenerics:settingsColumns]{settingsColumns()}}, \code{\link[omopgenerics:splitAdditional]{splitAdditional()}}, \code{\link[omopgenerics:splitAll]{splitAll()}}, \code{\link[omopgenerics:splitGroup]{splitGroup()}}, \code{\link[omopgenerics:splitStrata]{splitStrata()}}, \code{\link[omopgenerics:strataColumns]{strataColumns()}}, \code{\link[omopgenerics:suppress]{suppress()}}, \code{\link[omopgenerics:tidy]{tidy}}, \code{\link[omopgenerics:tidyColumns]{tidyColumns()}}, \code{\link[omopgenerics:uniteAdditional]{uniteAdditional()}}, \code{\link[omopgenerics:uniteGroup]{uniteGroup()}}, \code{\link[omopgenerics:uniteStrata]{uniteStrata()}}}
}}

