% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{setGlobalPlotOptions}
\alias{setGlobalPlotOptions}
\title{Set format options for all subsequent plots}
\usage{
setGlobalPlotOptions(style = NULL, type = NULL)
}
\arguments{
\item{style}{Character indicating which style to apply. Options are either
NULL (default ggplot style), or one of our pre-defined styles (refer to the
function \code{plotStyle()}).}

\item{type}{The desired format of the output plot. See \code{plotType()} for
supported plot types.}
}
\description{
Set format options for all subsequent plots unless state a different style in
a specific function
}
