% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanHeader.R
\name{spanHeader}
\alias{spanHeader}
\title{Format estimate_value column of summarised_result and compared_result
object.}
\usage{
spanHeader(result, header, delim = "\\n", includeHeaderName = TRUE)
}
\arguments{
\item{result}{A summarised_result or compared_result.}

\item{header}{Names of the columns to make headers. Names not corresponding
to a column of the table result, will be used as headers at the defined
position.}

\item{delim}{Delimiter to use to separate headers.}

\item{includeHeaderName}{Wheather to include the column name as header.}
}
\description{
Format estimate_value column of summarised_result and compared_result
object.
}
\examples{
\donttest{
result <- mockSummarisedResult()

result |>
  spanHeader(header = c("Study cohorts", "group_level", "Study strata",
                        "strata_name", "strata_level"),
             includeHeaderName = FALSE)
}
}
