% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitAdditional}
\alias{splitAdditional}
\title{Split additional_name and additional_level into the columns.}
\usage{
splitAdditional(result, overall = FALSE)
}
\arguments{
\item{result}{Omop result object (summarised_result or compared_result).}

\item{overall}{Whether to keep overall column if present.}
}
\description{
Split additional_name and additional_level into the columns.
}
\examples{
\donttest{
#library(visOmop)

mockSummarisedResult() |>
  splitAdditional()
}

}
