% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitAll}
\alias{splitAll}
\title{Split group, strata and additional into their respective columns.}
\usage{
splitAll(result)
}
\arguments{
\item{result}{Omop result object (summarised_result or compared_result).}
}
\value{
A dataframe with group, strata and additional name as columns.
}
\description{
Pivots the input dataframe so group, strata and additional name columns are
transformed into column names with values from the corresponding level columns
(group, strata, and additional).
}
\examples{
\donttest{
mockSummarisedResult() |>
  splitAll()
}

}
