% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitGroup}
\alias{splitGroup}
\title{Split group_name and group_level into columns.}
\usage{
splitGroup(result, overall = FALSE)
}
\arguments{
\item{result}{omop_result object.}

\item{overall}{Whether to keep overall column if present.}
}
\description{
Split group_name and group_level into columns.
}
\examples{
\donttest{
mockSummarisedResult() |>
  splitGroup()
}

}
