% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitNameLevel}
\alias{splitNameLevel}
\title{Split strata_name and strata_level into the columns.}
\usage{
splitNameLevel(
  result,
  name = "group_name",
  level = "group_level",
  keep = FALSE,
  overall = FALSE
)
}
\arguments{
\item{result}{Omop result object (summarised_result or compared_result).}

\item{name}{Column with the names.}

\item{level}{Column with the levels.}

\item{keep}{Whether to keep the original group_name and group_level columns.}

\item{overall}{Whether to keep overall column if present.}
}
\description{
Split strata_name and strata_level into the columns.
}
\examples{
\donttest{
#library(visOmop)

mockSummarisedResult() |>
  splitNameLevel()
}

}
