% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitStrata}
\alias{splitStrata}
\title{Split strata_name and strata_level into the columns.}
\usage{
splitStrata(result, keep = FALSE, overall = FALSE)
}
\arguments{
\item{result}{Omop result object (summarised_result or compared_result).}

\item{keep}{Whether to keep the original group_name and group_level columns.}

\item{overall}{Whether to keep overall column if present.}
}
\description{
Split strata_name and strata_level into the columns.
}
\examples{
\donttest{
#library(visOmop)

mockSummarisedResult() |>
  splitStrata()
}

}
