% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{strataColumns}
\alias{strataColumns}
\title{Identify strata columns in an omop result object.}
\usage{
strataColumns(result, overall = FALSE)
}
\arguments{
\item{result}{An omop result object.}

\item{overall}{Whether to keep overall column if present.}
}
\description{
Identify strata columns in an omop result object.
}
\examples{
\donttest{
mockSummarisedResult() |>
  strataColumns()
}

}
