% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{tableColumns}
\alias{tableColumns}
\title{Columns for the table functions}
\usage{
tableColumns(result)
}
\arguments{
\item{result}{A \verb{<summarised_result>} object.}
}
\value{
A character vector of supported columns for tables.
}
\description{
Names of the columns that can be used in the input arguments for the table
functions.
}
\examples{
result <- mockSummarisedResult()
tableColumns(result)

}
