% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{tableOptions}
\alias{tableOptions}
\title{Additional table formatting options for \code{visOmopTable()} and \code{visTable()}}
\usage{
tableOptions()
}
\value{
A named list of default options for table customisation.
}
\description{
This function provides a list of allowed inputs for the \code{.option} argument in
\code{visOmopTable()} and \code{visTable()}, and their corresponding default values.
}
\examples{
tableOptions()

}
