% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{tableStyleCode}
\alias{tableStyleCode}
\title{Supported predefined style codes for formatted tables}
\usage{
tableStyleCode(type = "gt", style = "default")
}
\arguments{
\item{type}{Character string specifying the formatted table class.
See \code{tableType()} for supported classes. Default is "gt".}

\item{style}{Supported predefined styles. Check which ones with the function
\code{tableStyle()}.}
}
\value{
The code expression to style the table for as specific pre-defined
style and table type.
}
\description{
Supported predefined style codes for formatted tables
}
\examples{
tableStyleCode("gt")
tableStyleCode("flextable")

}
