% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{tableType}
\alias{tableType}
\title{Supported table classes}
\usage{
tableType()
}
\value{
A character vector of supported table types.
}
\description{
This function returns the supported table classes that can be used in the
\code{type} argument of \code{visOmopTable()}, \code{visTable()}, and \code{formatTable()}
functions.
}
\examples{
tableType()

}
