% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{tidyColumns}
\alias{tidyColumns}
\title{Identify tidy columns of a summarised_result}
\usage{
tidyColumns(result)
}
\arguments{
\item{result}{A summarised_result.}
}
\value{
Table columns after applying \code{tidy()} function to a summarisd result.
}
\description{
Identifies and returns the columns that the tidy version of the
summarised_result will have.
}
\examples{
mockSummarisedResult() |>
  tidyColumns()

}
