% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidySummarisedResult.R
\name{tidySummarisedResult}
\alias{tidySummarisedResult}
\title{Formats the estimate_value column}
\usage{
tidySummarisedResult(
  result,
  split = TRUE,
  settings = "pivot",
  pivotEstimates = TRUE
)
}
\arguments{
\item{result}{A summarised_result.}

\item{split}{Whether to split all name/level column pairs or not.}

\item{settings}{"pivot" to pivot settings as columns, "remove" to remove them
from the table, or "keep" to keep them in the table.}

\item{pivotEstimates}{TRUE for pivotting estimates as columns.}
}
\value{
A tibble.
}
\description{
Provides tools for obtaining for tidying a summarised_result object.
}
\examples{
result <- mockSummarisedResult()

result |> formatEstimateValue(decimals = 1)

result |> formatEstimateValue(decimals = c(integer = 0, numeric = 1))

result |>
  formatEstimateValue(decimals = c(numeric = 1, count = 0))
}
