% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{uniteAdditional}
\alias{uniteAdditional}
\title{Unite one or more columns in additional_name-additional_level format.}
\usage{
uniteAdditional(x, cols)
}
\arguments{
\item{x}{Tibble or data.frame.}

\item{cols}{Columns to aggregate.}
}
\value{
A Tibble with the new columns.
}
\description{
Unite one or more columns in additional_name-additional_level format.
}
