% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{uniteGroup}
\alias{uniteGroup}
\title{Unite one or more columns in group_name-group_level format.}
\usage{
uniteGroup(x, cols)
}
\arguments{
\item{x}{Tibble or data.frame.}

\item{cols}{Columns to aggregate.}
}
\value{
A Tibble with the new columns.
}
\description{
Unite one or more columns in group_name-group_level format.
}
\examples{
\donttest{
x <- dplyr::tibble(
  variable = "number subjects",
  value = c(10, 15, 40, 78),
  sex = c("Male", "Female", "Male", "Female"),
  age_group = c("<40", ">40", ">40", "<40")
)

x

x |>
  uniteGroup(c("sex", "age_group"))
}

}
