% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{uniteNameLevel}
\alias{uniteNameLevel}
\title{Unite one or more columns in name-level format}
\usage{
uniteNameLevel(
  x,
  cols = character(0),
  name = "group_name",
  level = "group_level",
  keep = FALSE,
  ignore = c(NA, "overall")
)
}
\arguments{
\item{x}{A dataframe.}

\item{cols}{Columns to aggregate.}

\item{name}{Column name of the \code{name} column.}

\item{level}{Column name of the \code{level} column.}

\item{keep}{Whether to keep the original columns.}

\item{ignore}{Level values to ignore.}
}
\value{
A tibble with the new columns.
}
\description{
Unites targeted table columns into a pair of name-level columns.
}
