% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{additionalColumns}
\alias{additionalColumns}
\title{Identify additional columns in an omop result object}
\usage{
additionalColumns(result, overall = FALSE)
}
\arguments{
\item{result}{A summarised_result.}

\item{overall}{Whether to keep overall column if present.}
}
\value{
Unique values of the additional name column.
}
\description{
Identifies and returns the unique values in additional_name
column.
}
\examples{
mockSummarisedResult() |>
  additionalColumns()

}
