% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{barPlot}
\alias{barPlot}
\title{Create a bar plot visualisation from a summarised result object}
\usage{
barPlot(result, x, y, facet = NULL, colour = NULL)
}
\arguments{
\item{result}{A summarised result object.}

\item{x}{Column or estimate name that is used as x variable.}

\item{y}{Column or estimate name that is used as y variable}

\item{facet}{Variables to facet by, a formula can be provided to specify
which variables should be used as rows and which ones as columns.}

\item{colour}{Columns to use to determine the colors.}
}
\value{
A plot object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\donttest{
result <- mockSummarisedResult() |> dplyr::filter(variable_name == "age")

barPlot(
  result = result,
  x = "cohort_name",
  y = "mean",
  facet = c("age_group", "sex"),
  colour = "sex")
}

}
