% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatEstimateName.R
\name{formatEstimateName}
\alias{formatEstimateName}
\title{Format estimate_value column of summarised_result and compared_result
object.}
\usage{
formatEstimateName(result, format, keepNotFormatted = TRUE)
}
\arguments{
\item{result}{A summarised_result or compared_result.}

\item{format}{Named list of estimate name's to join.}

\item{keepNotFormatted}{Whether to keep not formatted.}
}
\description{
Format estimate_value column of summarised_result and compared_result
object.
}
\examples{
\donttest{
x <- 1
}

}
