% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSummarisedResult.R
\name{formatSummarisedResult}
\alias{formatSummarisedResult}
\title{Create a nice table from a summarised_result object.}
\usage{
formatSummarisedResult(
  result,
  tableType = "gt",
  header = NULL,
  estimateNameFormat = NULL,
  style = "default",
  decimals = c(integer = 0, numeric = 2, percentage = 2, proportion = 3),
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  .options = list()
)
}
\arguments{
\item{result}{A summarised_result object.}

\item{tableType}{Type of table to format into (currently accepting "gt").}

\item{header}{Names of the columns to make headers. Names not corresponding
to a column of the table result, will be used as headers at the defined
position.}

\item{estimateNameFormat}{Named list of estimate name's to join, sorted by
computation order.}

\item{style}{Named list that specifies how to style the different parts of
the table. Accepted entries are: title, subtitle, header, header_name,
header_level, column_name, group_label, and body. If NULL, the table will
not be styled (deafult of the chosed tableType), and if "deafult" it will
be styled as OMOP's deafult.}

\item{decimals}{Number of decimals per estimate_type (integer, numeric,
percentage, proportion).}

\item{title}{Title of the table, or NULL for no title.}

\item{subtitle}{Subtitle of the table, or NULL for no subtitle.}

\item{caption}{Caption for the table, or NULL for no caption. Text in
markdown formatting style (e.g. \verb{*Your caption here*} for caption in
italics).}

\item{.options}{list of deafult options (includeHeaderName = TRUE,
keepNotFormatted = TRUE, decimalMark = ".", bigMark = ",", na = "-",
groupNameCol = NULL, groupNameAsColumn = FALSE, groupOrder = NULL).}
}
\value{
A formatted table from a summarisedResult dataframe.
}
\description{
Create a nice table from a summarised_result object.
}
\examples{
\donttest{
mockSummarisedResult() |>
  formatSummarisedResult(tableType = "gt",
                         header = c("Study strata","strata_name",
                                    "strata_level"),
                         estimateNameFormat = c("N(\%)" =
                                                  "count (percentage\%)"),
                         style = "default",
                         decimals = c(
                           integer = 0, numeric = 2, percentage = 1,
                           proportion = 3
                         ),
                         .options = list("groupNameCol" = "group_level",
                                         "groupNameAsColumn" = FALSE,
                                         "groupOrder" = c("cohort1",
                                                          "cohort2"),
                                         "includeHeaderName" = FALSE))
}


}
