% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{groupColumns}
\alias{groupColumns}
\title{Identify group columns in an omop result object.}
\usage{
groupColumns(result, overall = FALSE)
}
\arguments{
\item{result}{An omop result object.}

\item{overall}{Whether to keep overall column if present.}
}
\description{
Identify group columns in an omop result object.
}
\examples{
\donttest{
mockSummarisedResult() |>
  groupColumns()
}

}
