% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtResult-package.R
\docType{package}
\name{gtResult-package}
\alias{gtResult}
\alias{gtResult-package}
\title{gtResult: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://catalamarti.github.io/gtSummarisedResult/}
  \item Report bugs at \url{https://github.com/catalamarti/gtSummarisedResult/issues}
}

}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
