% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockResults.R
\name{mockSummarisedResult}
\alias{mockSummarisedResult}
\title{A summarised_result object filled with mock data}
\usage{
mockSummarisedResult(settings = FALSE)
}
\arguments{
\item{settings}{If TRUE settings will be appended.}
}
\value{
An object of the class summarised_result with mock data.
}
\description{
Creates an object of the class summarised_result with mock data
for illustration purposes.
}
\examples{
mockSummarisedResult()


}
