% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockResults.R
\name{mockSummarisedResult}
\alias{mockSummarisedResult}
\title{A \verb{<summarised_result>} object filled with mock data}
\usage{
mockSummarisedResult()
}
\value{
An object of the class \verb{<summarised_result>} with mock data.
}
\description{
Creates an object of the class \verb{<summarised_result>} with mock data
for illustration purposes.
}
\examples{
mockSummarisedResult()


}
