% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visOmopTable.R
\name{optionsVisOmopTable}
\alias{optionsVisOmopTable}
\title{Additional arguments for the function visOmopTable}
\usage{
optionsVisOmopTable()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in
visOmopTable and their given default value.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
{
optionsVisOmopTable()
}


}
