% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-omopgenerics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{splitGroup}
\alias{uniteGroup}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{omopgenerics}{\code{\link[omopgenerics]{splitGroup}}, \code{\link[omopgenerics]{uniteGroup}}}
}}

