% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{setGlobalPlotOptions}
\alias{setGlobalPlotOptions}
\title{Set format options for all subsequent plots}
\usage{
setGlobalPlotOptions(style = NULL, type = NULL)
}
\arguments{
\item{style}{Visual theme to apply. Character, or \code{NULL}.
If a character, this may be either the name of a built-in style
(see \code{plotStyle()}), or a path to a \code{.yml} file that
defines a custom style. If \code{NULL}, the function will use the
explicit default style, unless a global style option is
set (see \code{setGlobalPlotOptions()}), or a \verb{_brand.yml} file is present
(in that order).
Refer to the package vignette on styles to learn more.}

\item{type}{Character string indicating the output plot format.
See \code{plotType()} for the list of supported plot types. If \code{type = NULL},
the function will use the global setting defined via \code{setGlobalPlotOptions()}
(if available); otherwise, a standard \code{ggplot2} plot is produced by default.}
}
\description{
Set format options for all subsequent plots unless state a different style in
a specific function
}
