% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanHeader.R
\name{spanHeader}
\alias{spanHeader}
\title{Format estimate_value column of summarised_result and compared_result
object.}
\usage{
spanHeader(result, header, includeHeader = TRUE)
}
\arguments{
\item{result}{A summarised_result or compared_result.}

\item{header}{Names of the columns to make headers. If a name does not correspond to a column of the table result, it will be used as a header at the defined position.}

\item{includeHeader}{Wheather to include the column name as header.}
}
\description{
Format estimate_value column of summarised_result and compared_result
object.
}
\examples{
\donttest{
result <- mockSummarisedResult()

result |> formatEstimateValue(decimals = 1)

result |> formatEstimateValue(decimals = c(integer = 0, numeric = 1))
}
}
