% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitAll}
\alias{splitAll}
\title{Split group, strata and additional into their respective columns.}
\usage{
splitAll(result)
}
\arguments{
\item{result}{Omop result object (summarised_result or compared_result).}
}
\description{
Split group, strata and additional into their respective columns.
}
\examples{
\donttest{
mockSummarisedResult() |>
  splitAll()
}

}
