% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitGroup}
\alias{splitGroup}
\title{Split group_name and group_level columns}
\usage{
splitGroup(result, overall = FALSE)
}
\arguments{
\item{result}{A dataframe with at least the columns group_name and
group_level.}

\item{overall}{Whether to keep overall column if present.}
}
\value{
A dataframe.
}
\description{
Pivots the input dataframe so the values of the column group_name are
transformed into columns that contain values from the group_level column.
}
\examples{
mockSummarisedResult() |>
  splitGroup()

}
