% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitGroup}
\alias{splitGroup}
\title{Split group_name and group_level into the columns.}
\usage{
splitGroup(result, overall = FALSE)
}
\arguments{
\item{result}{Omop result object (summarised_result or compared_result).}

\item{overall}{Whether to keep overall column if present.}
}
\description{
Split group_name and group_level into the columns.
}
\examples{
\donttest{
#library(visOmop)

mockSummarisedResult() |>
  splitGroup()
}

}
