% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitGroup}
\alias{splitGroup}
\title{Split group_name and group_level columns}
\usage{
splitGroup(
  result,
  keep = FALSE,
  fill = "overall",
  overall = lifecycle::deprecated()
)
}
\arguments{
\item{result}{A dataframe with at least the columns group_name and
group_level.}

\item{keep}{Whether to keep the original group_name and group_level columns.}

\item{fill}{Optionally, a character that specifies what value should be
filled in with when missing.}

\item{overall}{deprecated.}
}
\value{
A dataframe.
}
\description{
Pivots the input dataframe so the values of the column group_name are
transformed into columns that contain values from the group_level column.
}
\examples{
mockSummarisedResult() |>
  splitGroup()

}
