% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitGroup}
\alias{splitGroup}
\title{Split group_name and group_level into columns.}
\usage{
splitGroup(result, overall = FALSE)
}
\arguments{
\item{result}{omop_result object.}

\item{overall}{Whether to keep overall column if present.}
}
\value{
A dataframe with group_name as columns
}
\description{
Pivots the input dataframe so the values of the column group_name are
tranformed into columns with values from the group_level column
}
\examples{
mockSummarisedResult() |>
  splitGroup()

}
