% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitStrata}
\alias{splitStrata}
\title{Split strata_name and strata_level columns}
\usage{
splitStrata(result, overall = FALSE)
}
\arguments{
\item{result}{A dataframe with at least the columns strata_name and
strata_level.}

\item{overall}{Whether to keep overall column if present.}
}
\value{
A dataframe.
}
\description{
Pivots the input dataframe so the values of the column strata_name are
transformed into columns that contain values from the strata_level column.
}
\examples{
mockSummarisedResult() |>
  splitStrata()

}
