% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitStrata}
\alias{splitStrata}
\title{Split strata_name and strata_level into columns.}
\usage{
splitStrata(result, overall = FALSE)
}
\arguments{
\item{result}{omop_result object.}

\item{overall}{Whether to keep overall column if present.}
}
\description{
Split strata_name and strata_level into columns.
}
\examples{
\donttest{
mockSummarisedResult() |>
  splitStrata()
}

}
