% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{strataColumns}
\alias{strataColumns}
\title{Identify variables in strata_name column}
\usage{
strataColumns(result)
}
\arguments{
\item{result}{A tibble.}
}
\value{
Unique values of the strata name column.
}
\description{
Identifies and returns the unique values in strata_name column.
}
\examples{
mockSummarisedResult() |>
  strataColumns()

}
