% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{strataColumns}
\alias{strataColumns}
\title{Identify strata columns in an omop result object.}
\usage{
strataColumns(result, overall = FALSE)
}
\arguments{
\item{result}{An omop result object.}

\item{overall}{Whether to keep overall column if present.}
}
\value{
Unique values of the strata name column.
}
\description{
Identifies and returns the unique values in strata_name column.
}
\examples{
mockSummarisedResult() |>
  strataColumns()

}
