% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{tableStyle}
\alias{tableStyle}
\title{Supported predefined styles for formatted tables}
\usage{
tableStyle(type = "gt", styleName = "default")
}
\arguments{
\item{type}{Character string specifying the formatted table class.
See \code{tableType()} for supported classes. Default is "gt".}

\item{styleName}{A character string specifying the style name. Currently, the
package supports only one predefined style: "default".}
}
\value{
A code expression for the selected style and table type.
}
\description{
Supported predefined styles for formatted tables
}
\examples{
tableStyle("gt")
tableStyle("flextable")

}
