% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{tableStyle}
\alias{tableStyle}
\title{Pre-defined styles are available for tables}
\usage{
tableStyle()
}
\value{
A character vector indicating the style names.
}
\description{
This function provides a list of pre-defined styles for tables that can then
be used in the \code{style} argument of the table functions.
}
\examples{
tableStyle()

}
