% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.summarised_result}
\alias{tidy.summarised_result}
\title{Get a tidy visualization of a summarised_result object}
\usage{
\method{tidy}{summarised_result}(
  x,
  splitGroup = TRUE,
  splitStrata = FALSE,
  splitAdditional = TRUE,
  pivotEstimatesBy = c("variable_name", "variable_level", "estimate_name"),
  nameStyle = NULL,
  ...
)
}
\arguments{
\item{x}{A summarised_result.}

\item{pivotEstimatesBy}{Names from which pivot wider the estimate values. If
NULL the table will not be pivotted.}

\item{nameStyle}{Name style (glue package specifications) to customize names
when pivotting estimates. If NULL standard tidyr::pivot_wider formatting will
be used.}

\item{...}{For compatibility (not used).}

\item{split}{Whether to split all name/level column pairs or not.}
}
\value{
A tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Provides tools for obtaining a tidy version of a summarised_result object. It
pivots into columns the settings (if any), gives the option to split
group, strata, and additional columns, and whether and how to piviot
estimates.
}
\examples{
result <- mockSummarisedResult()

result |> tidy()

}
