% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{tidyColumns}
\alias{tidyColumns}
\title{Identify tidy columns of a \verb{<summarised_result>}}
\usage{
tidyColumns(result)
}
\arguments{
\item{result}{A \verb{<summarised_result>}.}
}
\value{
Table columns after applying \code{tidy()} function to a
\verb{<summarised_result>}.
}
\description{
Identifies and returns the columns that the tidy version of the
\verb{<summarised_result>} will have.
}
\examples{
mockSummarisedResult() |>
  tidyColumns()

}
