% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidySummarisedResult}
\alias{tidySummarisedResult}
\title{Get a custom tidy visualization of a summarised_result object}
\usage{
tidySummarisedResult(
  result,
  splitGroup = TRUE,
  splitStrata = TRUE,
  splitAdditional = TRUE,
  settingsColumns = colnames(settings(result)),
  pivotEstimatesBy = "estimate_name",
  nameStyle = NULL
)
}
\arguments{
\item{result}{A summarised_result.}

\item{splitGroup}{If TRUE it will split the group name-level column pair.}

\item{splitStrata}{If TRUE it will split the group name-level column pair.}

\item{splitAdditional}{If TRUE it will split the group name-level column pair.}

\item{settingsColumns}{Settings to be added as columns, by default all
settings will be added. If NULL or empty character vector, no settings will
be added.}

\item{pivotEstimatesBy}{Names from which pivot wider the estimate values. If
NULL the table will not be pivotted.}

\item{nameStyle}{Name style (glue package specifications) to customise names
when pivotting estimates. If NULL standard tidyr::pivot_wider formatting will
be used.}
}
\value{
A tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Provides tools for obtaining a tidy version of a summarised_result object.
}
\examples{
{
result <- mockSummarisedResult()

result |> tidySummarisedResult()

result |>
  tidySummarisedResult(
    settings = character(),
    pivotEstimatesBy = c("variable_name", "variable_level", "estimate_name")
  )

result |>
  tidySummarisedResult(
    settings = character(),
    pivotEstimatesBy = c("variable_name", "variable_level", "estimate_name"),
    nameStyle = "{estimate_name}_{variable_name}_{variable_level}"
  )
}
}
