% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidySummarisedResult.R
\name{tidySummarisedResult}
\alias{tidySummarisedResult}
\title{Get a tidy visualization of a summarised_result object}
\usage{
tidySummarisedResult(
  result,
  split = TRUE,
  pivotSettings = TRUE,
  pivotEstimatesBy = NULL,
  estimateLabels = NULL
)
}
\arguments{
\item{result}{A summarised_result.}

\item{split}{Whether to split all name/level column pairs or not.}

\item{pivotSettings}{TRUE to pivot settings into columns, and FALSE to remove
them form the results.}

\item{pivotEstimatesBy}{Names from which pivot wider the estimate values. If
NULL the table will not be pivotted.}

\item{estimateLabels}{Name style (glue package specifications) to customize names
when pivotting estimates.}
}
\value{
A tibble.
}
\description{
\\ifelse{html}{\\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\\strong{[Experimental]}}
Provides tools for obtaining for tidying a summarised_result object.
}
\examples{
result <- mockSummarisedResult()

result |> tidySummarisedResult()
}
