% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{uniteStrata}
\alias{uniteStrata}
\title{Unite one or more columns in strata_name-strata_level format.}
\usage{
uniteStrata(x, cols)
}
\arguments{
\item{x}{Tibble or data.frame.}

\item{cols}{Columns to aggregate.}
}
\value{
A Tibble with the new columns.
}
\description{
Unite one or more columns in strata_name-strata_level format.
}
