% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisOmop-package.R
\docType{package}
\name{visOmopResults-package}
\alias{visOmopResults}
\alias{visOmopResults-package}
\title{visOmopResults: Graphs and Tables for OMOP Results}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides methods to transform omop_result objects into formatted tables and figures, facilitating the visualization of study results working with the Observational Medical Outcomes Partnership (OMOP) Common Data Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/visOmopResults/}
  \item Report bugs at \url{https://github.com/oxford-pharmacoepi/visOmopResults/issues}
}

}
\author{
\strong{Maintainer}: N<U+00FA>ria Mercad<U+00E9>-Besora \email{nuria.mercadebesora@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0006-7948-3747}{ORCID})

Authors:
\itemize{
  \item Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
}

}
\keyword{internal}
