% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisOmop-package.R
\docType{package}
\name{visOmopResults-package}
\alias{visOmopResults}
\alias{visOmopResults-package}
\title{visOmopResults: Format summarisedResult objects into gt and flextable objects}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/visOmopResults/}
  \item Report bugs at \url{https://github.com/oxford-pharmacoepi/visOmopResults/issues}
}

}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

Authors:
\itemize{
  \item Núria Mercadé-Besora \email{nuria.mercadebesora@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0006-7948-3747}{ORCID})
}

}
\keyword{internal}
