% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visOmopResults-package.R
\docType{package}
\name{visOmopResults-package}
\alias{visOmopResults}
\alias{visOmopResults-package}
\title{visOmopResults: Graphs and Tables for OMOP Results}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides methods to transform omop_result objects into formatted tables and figures, facilitating the visualisation of study results working with the Observational Medical Outcomes Partnership (OMOP) Common Data Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu.github.io/visOmopResults/}
  \item \url{https://github.com/darwin-eu/visOmopResults}
  \item Report bugs at \url{https://github.com/darwin-eu/visOmopResults/issues}
}

}
\author{
\strong{Maintainer}: Núria Mercadé-Besora \email{nuria.mercadebesora@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0006-7948-3747}{ORCID})

Authors:
\itemize{
  \item Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
}

Other contributors:
\itemize{
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0847-4855}{ORCID}) [contributor]
  \item Elin Rowlands \email{elin.rowlands@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0007-6629-4661}{ORCID}) [contributor]
  \item Marta Alcalde-Herraiz \email{marta.alcaldeherraiz@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0002-4405-1814}{ORCID}) [contributor]
  \item Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID}) [contributor]
}

}
\keyword{internal}
